Allow `assert_no_accessibility_violations` to accept block
There might be cases where you wish to see if a page raises any
violations. Simply visiting a page with violations will fail, but it
could be nice to declare that in the test itself.

One possible use-case could be visiting a page that requires no
user interactions. A controller level test might cover the markup, but
the system level test would be created for testing accessibility
violations.