Support executing with `:cuprite` and `:playwright` drivers
In an effort to support more than Selenium-based drivers, this change
modifies the underlying `axe.js` integration.

Prior to this change, this gem relied entirely on the [axe-core-api][]
dependency to execute the audit, collect the results, and raise any
violations as testing exceptions.

Unfortunately, that gem does not provide official support for drivers
other than Selenium ([axe-core-gems#243][]).

To account for that shortcoming, this commit changes the gem to
integrate with the `axe.js` JavaScript code more directly, including
changes that invoke the interface directly through a combination of
Capybara's [execute_script][], [evaluate_script][], and
[evaluate_async_script][].

In support of that integration, this commit also extracts the
`AxeAuditor` and `RaiseReporter` classes for the sake of abstracting the
auditing and reporting steps (and for future integrations with other
styles of audit reporting).

[axe-core-api]: https://github.com/dequelabs/axe-core-gems/tree/develop/packages/axe-core-api
[axe-core-gems#243]: https://github.com/dequelabs/axe-core-gems/issues/243
[execute_script]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Session:execute_script
[evaluate_script]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Session:evaluate_script
[evaluate_async_script]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Session:evaluate_async_script